Revert "speed up ci builds from 15 to <7 minutes" (#1639)
Reverted due to stuck ci
