+++ .github/workflows/issue_commands.yml
@@ -9,7 +9,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout Actions
-        uses: actions/checkout@v3
+        uses: actions/checkout@v2
         with:
           repository: "grafana/grafana-github-actions"
           path: ./actions
