+++ engine/wait_for_test_mysql_start.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-until nc -z -v -w30 "${DATABASE_HOST:=mysql_test}" 3306
+until nc -z -v -w30 mysql_test 3306
 do
     echo "Waiting for database connection..."
     sleep 1
