+++ grafana-plugin/integration-tests/alerts/onCallSchedule.test.ts
@@ -1,9 +1,14 @@
 import { test } from '@playwright/test';
+import { configureOnCallPlugin } from '../utils/configurePlugin';
 import { verifyThatAlertGroupIsTriggered } from '../utils/alertGroup';
 import { createEscalationChain, EscalationStep } from '../utils/escalationChain';
 import { generateRandomValue } from '../utils/forms';
 import { createIntegrationAndSendDemoAlert } from '../utils/integrations';
 import { createOnCallSchedule } from '../utils/schedule';
+
+test.beforeEach(async ({ page }) => {
+  await configureOnCallPlugin(page);
+});
 
 test('we can create an oncall schedule + receive an alert', async ({ page }) => {
   const escalationChainName = generateRandomValue();
