+++ grafana-plugin/integration-tests/alerts/sms.test.ts
@@ -1,10 +1,15 @@
 import { test, expect } from '@playwright/test';
+import { configureOnCallPlugin } from '../utils/configurePlugin';
 import { GRAFANA_USERNAME } from '../utils/constants';
 import { createEscalationChain, EscalationStep } from '../utils/escalationChain';
 import { generateRandomValue } from '../utils/forms';
 import { createIntegrationAndSendDemoAlert } from '../utils/integrations';
 import { waitForSms } from '../utils/phone';
 import { configureUserNotificationSettings, verifyUserPhoneNumber } from '../utils/userSettings';
+
+test.beforeEach(async ({ page }) => {
+  await configureOnCallPlugin(page);
+});
 
 // TODO: enable once we've signed up for a MailSlurp account to receieve SMSes
 test.skip('we can verify our phone number + receive an SMS alert', async ({ page }) => {
