+++ grafana-plugin/integration-tests/escalationChains/searching.test.ts
@@ -1,6 +1,11 @@
 import { test, expect, Page } from '@playwright/test';
+import { configureOnCallPlugin } from '../utils/configurePlugin';
 import { generateRandomValue } from '../utils/forms';
 import { createEscalationChain } from '../utils/escalationChain';
+
+test.beforeEach(async ({ page }) => {
+  await configureOnCallPlugin(page);
+});
 
 const assertEscalationChainSearchWorks = async (
   page: Page,
@@ -15,18 +20,18 @@ const assertEscalationChainSearchWorks = async (
   await expect(page.getByTestId('escalation-chains-list')).toHaveText(escalationChainFullName);
 };
 
-// TODO: add tests for the new filtering. Commented out as this search doesn't exist anymore
-test.skip('searching allows case-insensitive partial matches', async ({ page }) => {
+test('searching allows case-insensitive partial matches', async ({ page }) => {
   const escalationChainName = `${generateRandomValue()} ${generateRandomValue()}`;
-  const [firstHalf, secondHalf] = escalationChainName.split(' ');
+  // const [firstHalf, secondHalf] = escalationChainName.split(' ');
 
   await createEscalationChain(page, escalationChainName);
 
-  await assertEscalationChainSearchWorks(page, firstHalf, escalationChainName);
-  await assertEscalationChainSearchWorks(page, firstHalf.toUpperCase(), escalationChainName);
-  await assertEscalationChainSearchWorks(page, firstHalf.toLowerCase(), escalationChainName);
-
-  await assertEscalationChainSearchWorks(page, secondHalf, escalationChainName);
-  await assertEscalationChainSearchWorks(page, secondHalf.toUpperCase(), escalationChainName);
-  await assertEscalationChainSearchWorks(page, secondHalf.toLowerCase(), escalationChainName);
+  // Commented as this search doesn't exist anymore TODO: add tests for the new filtering
+  // await assertEscalationChainSearchWorks(page, firstHalf, escalationChainName);
+  // await assertEscalationChainSearchWorks(page, firstHalf.toUpperCase(), escalationChainName);
+  // await assertEscalationChainSearchWorks(page, firstHalf.toLowerCase(), escalationChainName);
+  //
+  // await assertEscalationChainSearchWorks(page, secondHalf, escalationChainName);
+  // await assertEscalationChainSearchWorks(page, secondHalf.toUpperCase(), escalationChainName);
+  // await assertEscalationChainSearchWorks(page, secondHalf.toLowerCase(), escalationChainName);
 });
