+++ grafana-plugin/integration-tests/integrations/uniqueIntegrationNames.test.ts
@@ -1,5 +1,10 @@
 import { test, expect } from '@playwright/test';
+import { configureOnCallPlugin } from '../utils/configurePlugin';
 import { openCreateIntegrationModal } from '../utils/integrations';
+
+test.beforeEach(async ({ page }) => {
+  await configureOnCallPlugin(page);
+});
 
 test('integrations have unique names', async ({ page }) => {
   await openCreateIntegrationModal(page);
