+++ grafana-plugin/integration-tests/schedules/addOverride.test.ts
@@ -1,7 +1,12 @@
 import { test, expect } from '@playwright/test';
+import { configureOnCallPlugin } from '../utils/configurePlugin';
 import { clickButton, generateRandomValue } from '../utils/forms';
 import { createOnCallSchedule, getOverrideFormDateInputs } from '../utils/schedule';
 import dayjs from 'dayjs';
+
+test.beforeEach(async ({ page }) => {
+  await configureOnCallPlugin(page);
+});
 
 test('default dates in override creation modal are correct', async ({ page }) => {
   const onCallScheduleName = generateRandomValue();
