+++ grafana-plugin/integration-tests/utils/escalationChain.ts
@@ -1,4 +1,4 @@
-import { expect, Page } from '@playwright/test';
+import { Page } from '@playwright/test';
 
 import { clickButton, fillInInput, selectDropdownValue } from './forms';
 import { goToOnCallPage } from './navigation';
@@ -22,14 +22,6 @@ export const createEscalationChain = async (
   // go to the escalation chains page
   await goToOnCallPage(page, 'escalations');
 
-  /**
-   * wait for Esclation Chains page to fully load. this is because this can change which "New Escalation Chain"
-   * button is present
-   * ie. the one on the left hand side in the list vs the one in the center when no escalation chains exist
-   */
-  await page.getByTestId('page-title').locator('text=Escalation Chains').waitFor({ state: 'visible' });
-  await page.locator('text=Loading...').waitFor({ state: 'detached' });
-
   // open the create escalation chain modal
   (await page.waitForSelector('text=New Escalation Chain')).click();
 
@@ -38,7 +30,7 @@ export const createEscalationChain = async (
 
   // submit the form and wait for it to be created
   await clickButton({ page, buttonText: 'Create' });
-  await expect(page.getByTestId('escalation-chain-name')).toHaveText(escalationChainName);
+  await page.waitForSelector(`text=${escalationChainName}`);
 
   if (!escalationStep || !escalationStepValue) {
     return;
