+++ grafana-plugin/integration-tests/utils/forms.ts
@@ -54,7 +54,7 @@ const openSelect = async ({
   placeholderText,
   selectType,
   startingLocator,
-}: SelectDropdownValueArgs): Promise<Locator> => {
+}: SelectDropdownValueArgs): Promise<void> => {
   /**
    * we currently mix three different dropdown components in the UI..
    * so we need to support all of them :(
@@ -73,8 +73,6 @@ const openSelect = async ({
   const selectElement: Locator = (startingLocator || page).locator(selector);
   await selectElement.waitFor({ state: 'visible' });
   await selectElement.click();
-
-  return selectElement;
 };
 
 /**
@@ -87,15 +85,7 @@ const chooseDropdownValue = async ({ page, value, optionExactMatch = true }: Sel
   page.locator(`div[id^="react-select-"][id$="-listbox"] >> ${textMatchSelector(optionExactMatch, value)}`).click();
 
 export const selectDropdownValue = async (args: SelectDropdownValueArgs): Promise<void> => {
-  const selectElement = await openSelect(args);
-
-  /**
-   * use the select search to filter down the options
-   * TODO: get rid of the slice when we fix the GSelect component..
-   * without slicing this would fire off an API request for every key-stroke
-   */
-  await selectElement.type(args.value.slice(0, 5));
-
+  await openSelect(args);
   await chooseDropdownValue(args);
 };
