+++ grafana-plugin/package.json
@@ -59,7 +59,7 @@
     "@grafana/eslint-config": "^5.0.0",
     "@grafana/toolkit": "^9.2.4",
     "@jest/globals": "^27.5.1",
-    "@playwright/test": "^1.32.0",
+    "@playwright/test": "^1.28.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "12",
     "@testing-library/user-event": "^14.4.3",
