+++ grafana-plugin/playwright.config.ts
@@ -14,8 +14,7 @@ const config: PlaywrightTestConfig = {
   testDir: './integration-tests',
   globalSetup: './integration-tests/globalSetup.ts',
   /* Maximum time one test can run for. */
-  // TODO: set this back to 60 when GSelect component is refactored
-  timeout: 90 * 1000,
+  timeout: 60 * 1000,
   expect: {
     /**
      * Maximum time expect() should wait for the condition to be met.
@@ -28,10 +27,8 @@ const config: PlaywrightTestConfig = {
   /* Fail the build on CI if you accidentally left test.only in the source code. */
   forbidOnly: !!process.env.CI,
   /* Retry on CI only */
-  retries: process.env.CI ? 1 : 0,
-  // TODO: when GSelect component is refactored, run using 3 workers
-  // locally use one worker, on CI use 3
-  // workers: process.env.CI ? 3 : 1,
+  retries: process.env.CI ? 2 : 0,
+  /* Opt out of parallel tests on CI. */
   workers: 1,
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
   reporter: 'html',
